requirements.txt
matplotlib==3.5.2
opencv-python==4.6.0.66
scipy==1.8.1
Pillow==9.2.0
keras==2.9.0
numpy==1.21.4
pickle-mixin==1.0.2
streamlit==1.2.0
streamlit-option-menu==0.3.2
protobuf==3.20.1